+++ src/kernel/string.h
@@ -3,7 +3,7 @@
 #include <types.h>
 
 void *memcpy(void *dest, const void *src, size_t count);
-void *memset(void *dest, u8 val, size_t count);
+void *memset(void *dest, i32 val, size_t count);
 void *memmove(void *dest, const void *src, size_t count);
 
 size_t strlen(const char *s);
